[mobile + web + libs/ui + libs/wm] testing wm impl with mobile + moving Home screen to shared libs/ui folder
