+++ apps/mobile/src/pages/Home.tsx
@@ -1,41 +1,11 @@
-import { H1, PrimaryButton } from '@end/components';
-import { useCallback, ComponentType } from 'react';
-import { Text } from 'tamagui';
-import { database } from '@end/wm/rn';
-import { Planet } from '@end/wm/core';
-import {
-  compose,
-  withDatabase,
-  withObservables,
-} from '@nozbe/watermelondb/react';
-import { Database, Query } from '@nozbe/watermelondb';
+import { Home as H, PrimaryButton } from '@end/components';
+import { database, sync } from '@end/wm/rn';
 
-function Home({ allPlanets }: { allPlanets: Planet[] }) {
-  const addPlanet = useCallback(async () => {
-    await database.write(async () => {
-      await database.get('planets').create((planet: any) => {
-        planet.name = Math.random().toString();
-      });
-    });
-  }, []);
-
+export default function Home({ logOut }: { logOut: () => void }) {
   return (
     <>
-      <H1>Home</H1>
-      {allPlanets.map((planet) => (
-        <Text>{planet.name}</Text>
-      ))}
-      <PrimaryButton onPress={addPlanet}>Add planet</PrimaryButton>
+      <H database={database} sync={sync} />
+      <PrimaryButton onPress={logOut}>Logout</PrimaryButton>
     </>
   );
 }
-
-export default compose(
-  withDatabase,
-  withObservables(
-    [],
-    ({ database }: { database: Database }): { allPlanets: Query<Planet> } => ({
-      allPlanets: database.get<Planet>('planets').query(),
-    })
-  ) as (arg0: unknown) => ComponentType
-)(Home) as ComponentType;
