+++ libs/ui/rnw/src/index.ts
@@ -1,5 +1,6 @@
 export * from './lib/components';
 export * from './lib/Landing';
+export * from './lib/Home';
 export * from './lib/Typography';
 export * from './lib/Layout';
 export * from './lib/Display';
